AI social media posting Agent/requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
logging-config>=1.0.0
